Revert "Fix mail sender"
