+++ controllers/userControllers/sendHelpMsg.js
@@ -4,27 +4,13 @@ export async function sendHelpMsg(req, res, next) {
   const { email, comment } = req.body;
 
   const helpEmail = {
-    Messages: [
-      {
-        From: {
-          Email: process.env.UKRNET_SENDER,
-        },
-        To: [
-          {
-            Email: process.env.SUPPORT_EMAIL,
-          },
-          {
-            Email: process.env.SUB_SUPPORT_EMAIL,
-          },
-        ],
-        Subject: "Need help",
-        HTMLPart: `
-            <div>${comment}</div></br>
-            <div><a href="mailto:${email}" target="_blank">${email}</a></div>`,
-        TEXTPart: `${comment} ${email}`,
-        CustomID: "AppGettingStartedTest",
-      },
-    ],
+    to: process.env.SUPPORT_EMAIL,
+    from: process.env.UKRNET_SENDER,
+    subject: "Need Help",
+    html: `
+      <div>${comment}</div></br>
+      <div><a href="mailto:${email}" target="_blank">${email}</a></div>`,
+    text: `${comment} ${email}`,
   };
 
   await sendEmail(helpEmail);
